requirements.txt
django>=4.0.0,<=4.2.0
djangorestframework
pyyaml
requests
django-cors-headers
autopep8
markdown
pygments
django-filter
django-guardian